_algorithms_challenges/pybites/bitesofpy-master/!101-200/118/duplicates.py
from collections import defaultdict


def get_duplicate_indices(words):
    """Given a list of words, loop through the words and check for each
       word if it occurs more than once.
       If so return the index of its first ocurrence.
       For example in the following list 'is' and 'it'
       occurr more than once, and they are at indices 0 and 1 so you would
       return [0, 1]:
       ['is', 'it', 'true', 'or', 'is', 'it', 'not?'] => [0, 1]
       Make sure the returning list is unique and sorted in ascending order."""
    wordlist = defaultdict(list)
    for pos, word in enumerate(words):
        wordlist[word].append(pos)
    return sorted(p[0] for p in wordlist.values() if len(p) > 1)
